+++ tests/e2e/utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+## Added
+
+- Update `shopper.addToCartFromShopPage()` to accept product Id or Title
+
 # 0.1.6
 
 ## Added